add compatibility with clangd for better editing with VSCode

+++ src/list_menu.c
@@ -1261,17 +1261,17 @@ void ListMenuSetUpRedOutlineCursorSpriteOamTable(u16 rowWidth, u16 rowHeight, st
     s32 i, j, id = 0;
 
     subsprites[id] = sSubsprite_RedOutline1;
-    subsprites[id].x = 136;
-    subsprites[id].y = 136;
+    subsprites[id].x = (s8) 136;
+    subsprites[id].y = (s8) 136;
     id++;
 
     subsprites[id] = sSubsprite_RedOutline2;
     subsprites[id].x = rowWidth + 128;
-    subsprites[id].y = 136;
+    subsprites[id].y = (s8) 136;
     id++;
 
     subsprites[id] = sSubsprite_RedOutline7;
-    subsprites[id].x = 136;
+    subsprites[id].x = (s8) 136;
     subsprites[id].y = rowHeight + 128;
     id++;
 
@@ -1301,7 +1301,7 @@ void ListMenuSetUpRedOutlineCursorSpriteOamTable(u16 rowWidth, u16 rowHeight, st
         for (j = 8; j < rowHeight - 8; j += 8)
         {
             subsprites[id] = sSubsprite_RedOutline4;
-            subsprites[id].x = 136;
+            subsprites[id].x = (s8) 136;
             subsprites[id].y = j - 120;
             id++;
